requirements.txt
timm==1.0.7
fastcore
fastai
scikit-learn
Pillow
uvicorn
fastapi
pytest